Rename firmware to fw_updater
